[filament_view] Fixed transform setup for nested objects
When models are imported, their child renderables need to have their position set in our systems for the colliders to work properly. This was a bit tricky because it turns out the GLTF loader only provides us with a composed matrix, so that needed to be reversed.
Also removed most vector/matrix utils implementations, turns out Filament already had them